chore: move Api.js to a new utils directory and organize index.js
