+++ src/pages/index.js
@@ -6,46 +6,9 @@ import {
   settings,
 } from "../scripts/validation.js";
 
-import API from "../scripts/Api.js";
+import API from "../utils/Api.js";
 
-// Selects Edit profile modal and buttons
-const editModalOpenBtn = document.querySelector(".profile__edit-button");
-const editProfileModal = document.querySelector("#edit-profile-modal");
-const editModalCloseBtn = editProfileModal.querySelector(
-  ".modal__button-close"
-);
-const editProfileSubmitBtn = editProfileModal.querySelector(
-  ".modal__button-submit"
-);
-
-// Selects the <ul> element from the html to accept the newly created cards
-const cardsList = document.querySelector(".cards__list");
-
-// Selects New post modal and buttons
-const newPostModal = document.querySelector("#new-post-modal");
-const newPostOpenBtn = document.querySelector(".profile__button-add");
-const newPostCloseBtn = newPostModal.querySelector(".modal__button-close");
-const newPostSubmitBtn = newPostModal.querySelector(".modal__button-submit");
-
-// Selects profile section elements
-const editAvatarOpenBtn = document.querySelector(".profile__avatar-button");
-const editAvatarModal = document.querySelector("#avatar-modal");
-const editAvatarCloseBtn = editAvatarModal.querySelector(
-  ".modal__button-close"
-);
-const avatarInput = editAvatarModal.querySelector("#image-input");
-const editAvatarSubmitBtn = editAvatarModal.querySelector(
-  ".modal__button-submit"
-);
-const profileAvatar = document.getElementById("profile-avatar");
-const profileName = document.querySelector(".profile__name");
-const nameInput = editProfileModal.querySelector("#name-input");
-const profileDescription = document.querySelector(".profile__description");
-const descriptionInput = editProfileModal.querySelector("#description-input");
-
-// Selects new post inputs
-const linkInput = newPostModal.querySelector("#image-input");
-const captionInput = newPostModal.querySelector("#caption-input");
+//VARIABLES
 
 // Selects each form
 const editProfileForm = document.forms["edit-profile"];
@@ -53,11 +16,51 @@ const newPostForm = document.forms["new-post"];
 const cardDeleteForm = document.forms["delete-card"];
 const newAvatarForm = document.forms["new-avatar"];
 
-// Selects Preview modal elements
+// Selects Edit profile modal and buttons
+const editProfileModal = document.querySelector("#edit-profile-modal");
+const editModalOpenBtn = document.querySelector(".profile__edit-button");
+const editModalCloseBtn = editProfileModal.querySelector(
+  ".modal__button-close"
+);
+const editProfileSubmitBtn = editProfileModal.querySelector(
+  ".modal__button-submit"
+);
+
+//EditProfileForm inputs and elements
+const nameInput = editProfileModal.querySelector("#name-input");
+const descriptionInput = editProfileModal.querySelector("#description-input");
+const profileName = document.querySelector(".profile__name");
+const profileDescription = document.querySelector(".profile__description");
+
+// Selects New post modal and buttons
+const newPostModal = document.querySelector("#new-post-modal");
+const newPostOpenBtn = document.querySelector(".profile__button-add");
+const newPostCloseBtn = newPostModal.querySelector(".modal__button-close");
+const newPostSubmitBtn = newPostModal.querySelector(".modal__button-submit");
+
+// Selects new post inputs
+const linkInput = newPostModal.querySelector("#image-input");
+const captionInput = newPostModal.querySelector("#caption-input");
+
+// Selects Preview modal, buttons, and elements
 const previewModal = document.querySelector("#preview-modal");
 const previewCloseBtn = previewModal.querySelector(".modal__button-close");
 const previewImage = previewModal.querySelector(".modal__img");
 const previewCaption = previewModal.querySelector(".modal__caption");
+
+//Selects edit avatar modal and buttons
+const editAvatarModal = document.querySelector("#avatar-modal");
+const editAvatarOpenBtn = document.querySelector(".profile__avatar-button");
+const editAvatarCloseBtn = editAvatarModal.querySelector(
+  ".modal__button-close"
+);
+const editAvatarSubmitBtn = editAvatarModal.querySelector(
+  ".modal__button-submit"
+);
+
+//Selects avatar element and input
+const profileAvatar = document.getElementById("profile-avatar");
+const avatarInput = editAvatarModal.querySelector("#image-input");
 
 //Selects delete modal and buttons
 const deleteCardModal = document.querySelector("#delete-card-modal");
@@ -67,45 +70,23 @@ const deleteCardModalCloseBtn = deleteCardModal.querySelector(
 const cancelBtn = deleteCardModal.querySelector("#cancel");
 const deleteSubmitBtn = deleteCardModal.querySelector(".modal__button-submit");
 
-const api = new API({
-  baseUrl: "https://around-api.en.tripleten-services.com/v1",
-  headers: {
-    authorization: "640264a7-d833-4076-a643-c71cc77da761",
-    "content-type": "application/json",
-  },
-});
+// Selects the <ul> element from the html to accept the newly created cards
+const cardsList = document.querySelector(".cards__list");
 
-//Make GET requests to retrieve data upon page load
-api
-  .getAppData()
-  .then((data) => {
-    //data is an array with two elements that comes froms a Promise.all call that
-    // loads the initial data from the server. The first element is an array of cards
-    //and the second element is the userInfo.
-    const [cardData, userData] = data;
-    cardData.forEach((item) => {
-      const card = getCardElement(item);
-      cardsList.append(card);
-    });
-
-    profileAvatar.src = userData.avatar;
-    profileName.textContent = userData.name;
-    profileDescription.textContent = userData.about;
-  })
-  .catch((error) => console.error(error));
-
-enableValidation(settings);
+//FUNCTIONS
 
 // General purpose open and close modal logic w/ Escape key close
 function openModal(modal) {
   modal.classList.add("modal_is-opened");
   document.addEventListener("keydown", handleKeyDown);
 }
+
 function closeModal(modal) {
   modal.classList.remove("modal_is-opened");
   //Removed keydown event listener for performance reasons
   document.removeEventListener("keydown", handleKeyDown);
 }
+
 // click to close logic
 const handleClickToClose = (modal) => {
   modal.addEventListener("click", (e) => {
@@ -114,6 +95,7 @@ const handleClickToClose = (modal) => {
     }
   });
 };
+
 //Handle keydown event Escape close
 function handleKeyDown(e) {
   modalList.forEach((modal) => {
@@ -122,46 +104,6 @@ function handleKeyDown(e) {
     }
   });
 }
-//handle click to close events
-const modalList = document.querySelectorAll(".modal");
-modalList.forEach((modal) => handleClickToClose(modal));
-
-// New post open event listener
-newPostOpenBtn.addEventListener("click", () => {
-  handleSubmitBtnUI(newPostSubmitBtn);
-  openModal(newPostModal);
-});
-// Edit Profile open event listener
-editModalOpenBtn.addEventListener("click", function () {
-  handleSubmitBtnUI(editProfileSubmitBtn);
-  openModal(editProfileModal);
-  resetValidation(editProfileForm, [nameInput, descriptionInput], settings);
-  fillEditProfileForm();
-});
-
-// New post submisson handler
-function handleAddCardSubmit(evt) {
-  evt.preventDefault();
-  const inputData = {
-    name: captionInput.value,
-    link: linkInput.value,
-  };
-  api
-    .addNewCard(inputData)
-    .then((data) => {
-      handleSubmitBtnUI(newPostSubmitBtn);
-      const card = getCardElement(data);
-      cardsList.prepend(card);
-      newPostForm.reset();
-      closeModal(newPostModal);
-      disableSubmitButton(newPostSubmitBtn, settings);
-    })
-    // figure out how to make an error message appear in the form
-    .catch((error) => console.error(error));
-}
-
-// New post submit listener
-newPostForm.addEventListener("submit", handleAddCardSubmit);
 
 // Sets default values when the edit modal form is opened.
 function fillEditProfileForm() {
@@ -170,64 +112,11 @@ function fillEditProfileForm() {
   descriptionInput.value = profileDescription.textContent;
 }
 
-// Edit Profile submission handler
-function handleEditProfileFormSubmit(e) {
-  // Prevents the page from refreshing when the submit button is clicked
-  e.preventDefault();
-  const inputData = {
-    name: nameInput.value,
-    about: descriptionInput.value,
-  };
-  api
-    .editUserInfo(inputData)
-    .then((data) => {
-      handleSubmitBtnUI(editProfileSubmitBtn);
-      //If the promise resolves, the DOM updates with the new Values saved to the server
-      profileName.textContent = data.name;
-      profileDescription.textContent = data.about;
-      closeModal(editProfileModal);
-      disableSubmitButton(editProfileSubmitBtn, settings);
-    })
-    //Make a visible UX response later
-    .catch((error) => console.error(error));
-}
-
-// Edit Profile submit listener
-editProfileForm.addEventListener("submit", handleEditProfileFormSubmit);
-
-//Close button listeners
-editModalCloseBtn.addEventListener("click", () => closeModal(editProfileModal));
-newPostCloseBtn.addEventListener("click", () => closeModal(newPostModal));
-previewCloseBtn.addEventListener("click", () => closeModal(previewModal));
-deleteCardModalCloseBtn.addEventListener("click", () =>
-  closeModal(deleteCardModal)
-);
-cancelBtn.addEventListener("click", () => closeModal(deleteCardModal));
-editAvatarCloseBtn.addEventListener("click", () => closeModal(editAvatarModal));
-
-//Delete card handler
-let selectedCard;
-let selectedCardId;
-function handleDeleteCard(cardELement, data) {
-  selectedCard = cardELement;
-  selectedCardId = data._id;
-  deleteSubmitBtn.textContent = "Delete";
-  openModal(deleteCardModal);
-}
-function handleDeleteCardSubmit(selectedCard, selectedCardId) {
-  api
-    .removeCard(selectedCardId)
-    .then(() => {
-      deleteSubmitBtn.textContent = "Deleting";
-      selectedCard.remove();
-      closeModal(deleteCardModal);
-    })
-    .catch((error) => console.error(error));
-}
-cardDeleteForm.addEventListener("submit", (e) => {
-  e.preventDefault();
-  handleDeleteCardSubmit(selectedCard, selectedCardId);
-});
+const handleCardLikes = (isLiked, cardLikeBtn) => {
+  isLiked
+    ? cardLikeBtn.classList.add("card__like-button_active")
+    : cardLikeBtn.classList.remove("card__like-button_active");
+};
 
 function handleSubmitBtnUI(btn) {
   if (btn.disabled === true) {
@@ -237,34 +126,13 @@ function handleSubmitBtnUI(btn) {
   }
 }
 
-//edit avatar
-editAvatarOpenBtn.addEventListener("click", (e) => {
-  handleSubmitBtnUI(editAvatarSubmitBtn);
-  openModal(editAvatarModal);
-});
-
-function handleNewAvatarSubmit() {
-  api
-    .editAvatar(avatarInput.value)
-    .then((data) => {
-      handleSubmitBtnUI(editAvatarSubmitBtn);
-      profileAvatar.src = data.avatar;
-      closeModal(editAvatarModal);
-      disableSubmitButton(editAvatarSubmitBtn, settings);
-    })
-    .catch((error) => console.error(error));
+//Attatched to the trashcan btn upon card creation
+function handleDeleteCard(cardELement, data) {
+  selectedCard = cardELement;
+  selectedCardId = data._id;
+  deleteSubmitBtn.textContent = "Delete";
+  openModal(deleteCardModal);
 }
-
-newAvatarForm.addEventListener("submit", (e) => {
-  e.preventDefault();
-  handleNewAvatarSubmit();
-});
-
-const handleCardLikes = (isLiked, cardLikeBtn) => {
-  isLiked
-    ? cardLikeBtn.classList.add("card__like-button_active")
-    : cardLikeBtn.classList.remove("card__like-button_active");
-};
 
 // Creates cards
 function getCardElement(data) {
@@ -325,6 +193,163 @@ function getCardElement(data) {
   return cardElement;
 }
 // End of getCardElement function
+
+//FORM SUBMISSION HANDLERS
+
+function handleDeleteCardSubmit(selectedCard, selectedCardId) {
+  api
+    .removeCard(selectedCardId)
+    .then(() => {
+      deleteSubmitBtn.textContent = "Deleting";
+      selectedCard.remove();
+      closeModal(deleteCardModal);
+    })
+    .catch((error) => console.error(error));
+}
+
+function handleNewAvatarSubmit() {
+  api
+    .editAvatar(avatarInput.value)
+    .then((data) => {
+      handleSubmitBtnUI(editAvatarSubmitBtn);
+      profileAvatar.src = data.avatar;
+      closeModal(editAvatarModal);
+      disableSubmitButton(editAvatarSubmitBtn, settings);
+    })
+    .catch((error) => console.error(error));
+}
+
+function handleEditProfileFormSubmit(e) {
+  // Prevents the page from refreshing when the submit button is clicked
+  e.preventDefault();
+  const inputData = {
+    name: nameInput.value,
+    about: descriptionInput.value,
+  };
+  api
+    .editUserInfo(inputData)
+    .then((data) => {
+      handleSubmitBtnUI(editProfileSubmitBtn);
+      //If the promise resolves, the DOM updates with the new Values saved to the server
+      profileName.textContent = data.name;
+      profileDescription.textContent = data.about;
+      closeModal(editProfileModal);
+      disableSubmitButton(editProfileSubmitBtn, settings);
+    })
+    //Make a visible UX response later
+    .catch((error) => console.error(error));
+}
+
+// New post submisson handler
+function handleAddCardSubmit(evt) {
+  evt.preventDefault();
+  const inputData = {
+    name: captionInput.value,
+    link: linkInput.value,
+  };
+  api
+    .addNewCard(inputData)
+    .then((data) => {
+      handleSubmitBtnUI(newPostSubmitBtn);
+      const card = getCardElement(data);
+      cardsList.prepend(card);
+      newPostForm.reset();
+      closeModal(newPostModal);
+      disableSubmitButton(newPostSubmitBtn, settings);
+    })
+    // figure out how to make an error message appear in the form
+    .catch((error) => console.error(error));
+}
+
+//FLOW
+
+const api = new API({
+  baseUrl: "https://around-api.en.tripleten-services.com/v1",
+  headers: {
+    authorization: "640264a7-d833-4076-a643-c71cc77da761",
+    "content-type": "application/json",
+  },
+});
+
+//Make GET requests to retrieve data upon page load
+api
+  .getAppData()
+  .then((data) => {
+    //data is an array with two elements that comes froms a Promise.all call that
+    // loads the initial data from the server. The first element is an array of cards
+    //and the second element is the userInfo.
+    const [cardData, userData] = data;
+    cardData.forEach((item) => {
+      const card = getCardElement(item);
+      cardsList.append(card);
+    });
+
+    profileAvatar.src = userData.avatar;
+    profileName.textContent = userData.name;
+    profileDescription.textContent = userData.about;
+  })
+  .catch((error) => console.error(error));
+
+enableValidation(settings);
+
+// Open button listeners
+
+editAvatarOpenBtn.addEventListener("click", (e) => {
+  handleSubmitBtnUI(editAvatarSubmitBtn);
+  openModal(editAvatarModal);
+});
+
+// Edit Profile open event listener
+editModalOpenBtn.addEventListener("click", function () {
+  handleSubmitBtnUI(editProfileSubmitBtn);
+  openModal(editProfileModal);
+  resetValidation(editProfileForm, [nameInput, descriptionInput], settings);
+  fillEditProfileForm();
+});
+
+// New post open event listener
+newPostOpenBtn.addEventListener("click", () => {
+  handleSubmitBtnUI(newPostSubmitBtn);
+  openModal(newPostModal);
+});
+
+//Close button listeners
+editModalCloseBtn.addEventListener("click", () => closeModal(editProfileModal));
+newPostCloseBtn.addEventListener("click", () => closeModal(newPostModal));
+previewCloseBtn.addEventListener("click", () => closeModal(previewModal));
+deleteCardModalCloseBtn.addEventListener("click", () =>
+  closeModal(deleteCardModal)
+);
+cancelBtn.addEventListener("click", () => closeModal(deleteCardModal));
+editAvatarCloseBtn.addEventListener("click", () => closeModal(editAvatarModal));
+
+//handle click to close events
+const modalList = document.querySelectorAll(".modal");
+modalList.forEach((modal) => handleClickToClose(modal));
+
+//Form Submission listeners
+
+// New post submit listener
+newPostForm.addEventListener("submit", handleAddCardSubmit);
+
+// Edit Profile submit listener
+editProfileForm.addEventListener("submit", handleEditProfileFormSubmit);
+
+//For deleting cards
+//Remove card from the DOM
+let selectedCard;
+//Remove card from the server
+let selectedCardId;
+
+cardDeleteForm.addEventListener("submit", (e) => {
+  e.preventDefault();
+  handleDeleteCardSubmit(selectedCard, selectedCardId);
+});
+
+newAvatarForm.addEventListener("submit", (e) => {
+  e.preventDefault();
+  handleNewAvatarSubmit();
+});
 
 export {
   editModalOpenBtn,
